System.Data.DataTable seems to be not fully supported in WebGL builds, moving it to the XUnit tests for now
