+++ CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/MathTests.cs
@@ -38,6 +38,23 @@ namespace com.csutil.tests {
             Assert.Equal(1, Numbers.Calculate("0,5 + 0,5"));
         }
 
+        [Fact]
+        public static void TestDataTable() {
+            var dt = new System.Data.DataTable();
+            dt.Columns.Add("A", typeof(int));
+            dt.Columns.Add("B", typeof(int));
+            dt.Rows.Add(11, 12); // Insert a row with A=4, B=1
+            // Querying the table for specific entries:
+            var boolResult = dt.Select("A>B-2").Length > 0;
+            Assert.True(boolResult); // 11 > 12-2
+            // Add a result column that calculates a formula based on the entries:
+            var columnName = "Result Column";
+            dt.Columns.Add(columnName, typeof(int), "A+B*2");
+            var rowNr = 0;
+            var valResult = dt.Rows[rowNr][columnName];
+            Assert.Equal(35, valResult); // 11 + 12*2  = 35
+        }
+
     }
 
 }
